Changed all instances of post to thread
