+++ app.js
@@ -10,7 +10,7 @@ var sequelizeConnection = require("./config/sequelizeConnection");
 
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
-var postRouter = require('./routes/post');
+var threadRouter = require('./routes/thread');
 var commentsRouter = require('./routes/comments');
 
 var app = express();
@@ -53,7 +53,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
-app.use('/post', postRouter);
+app.use('/thread', threadRouter);
 app.use('/static', express.static('public'));
 
 // catch 404 and forward to error handler
