+++ routes/index.js
@@ -4,7 +4,7 @@ var router  = express.Router();
 /* GET home page. */
 router.get('/', function(req, res, next) {
   //Get all posts
-  res.render('index', { title: 'Express', 'posts':[]});
+  res.render('index', { title: 'Express', 'threads':[]});
 });
 
 module.exports = router;
