+++ views/base/layout.pug
@@ -6,8 +6,7 @@ html
     meta(name='description' content='')
     meta(name='author' content='')
     // Bootstrap core CSS
-    title Blog Post - Start Bootstrap Template
-    title= title
+    title Tagless
     block stylesheets
   body
     block content
