--- views/elements/profile_widget.pug
@@ -1,15 +1,0 @@
-.col-md-4
-	.card.my-4
-		.card-body
-			.image-block
-				img.className#profilePicture(src=imageURL 
-									width="79"
-									height="79"
-									onerror="this.onerror=null;this.src='https://www.redditstatic.com/avatars/avatar_default_11_FFD635.png';")
-
-			| #{username}
-
-			//hay que ver donde se pone el style
-			p(style="margin-top:0;margin-bottom:14px;font-size:12px;line-height:18px;color:#222;") #{description}
-
-			a.nav-link(href="profile/settings" id="urluploader") Settings
