+++ views/elements/thread_submit.pug
@@ -1,7 +1,7 @@
 div.card.my-4
   h5.card-header Make a thread:
   div.card-body
-    form(method='POST' action='/post/submit')
+    form(method='POST' action='/thread/submit')
     div.form-group
       input#title.form-control(type='text', placeholder='Title' name='title')
     div.form-group
